+++ src/utils/math/easings.ts
@@ -24,7 +24,7 @@ export const EasingsFn: {
    readonly [easing in Easings]: EasingFunction;
 } = {
    easeLinear: (x) => x,
-   easeStep: (x) => x,
+   easeStep: (x) => Math.floor(x),
    easeInQuad: (x) => Math.pow(x, 2),
    easeOutQuad: (x) => x * (2 - x),
    easeInOutQuad: (x) => (x < 0.5 ? 2 * x * x : -1 + (4 - 2 * x) * x),