+++ src/utils/misc/hasOwn.ts
@@ -1,0 +1,3 @@
+export function hasOwn(object: object, property: PropertyKey): boolean {
+   return Object.prototype.hasOwnProperty.call(object, property);
+}